Upload of intitial pass at GUI assets
These are drawn with javascript and canvas. Its all just an initial pass. I was hoping we could interchange files this way to help streamline the development of the GUI but I realize that this stuff may not be that portable.

I hope its clean enough to get the knobs going.