Final Fixes for CI Pipeline (#20)
Final fixes to make Rack build in the CI Pipeline on Windows
(which basically amount to "don't assume the C++17 filesystem
API is there; and g++ on windows is not the same as WINDOWS
in surge). So a set of directies in surge and make rules here and
we are good.