A Simple Working LFO module
The Surge LFO is powerful, triggerable, flexible, etc...
This commit implements it very mechanically, showing all
the parameters with a gross UI even if that doesn't make sense,
but the module works, ready for design. We also introduce a
set of utility classes to make binding to Surge parameters substantially
simpler and use them in the LFO. Finally we apply a couple of tiny
cleanups to the ADSR and to my wonky build scripts.

Closes #37.