Redo FX; Fix Presets (#132)
This diff compeltely redoes the way we do FX. Rather than one
monster FX module this makes a module-per-FX. In this commit
they all share the same utilitarian UI but that will obviously
change allowing things to be more compact on a module-by-module
basis. Additionally this change fixes an exception in the preset
load path on linux.