Hack-around a bug to load a WT (#156)
WaveTable loads of small table on restart are not working. This
hacks around that with a gross but effective hack of forcing a
large table load. We should really fux the underlying but in surge
that causes this. But what to we say to bugs? not today!

Addresses #124 which I should rename after committing this.