Path to cleaner parameters (#177)
This commit sets out a path to using a more consistent parmeter
binding class, and in the LFO where that conversion is complete
showing how that gives you the Rack UI features like smarter
right mouse editors and labels.

This addressed parts of #99, #100 and #147. I think I'm going to
collapse those all into #147 and use that to hang the rest of the
implementation of this in the other classes.