Add TempoSync to Rotary (#186)
Per #149 we need to finish the TempoSync implementation so here
I add it all to the Rotary Panel, which is cool. Then I find out
that Rotary has a bug that it advertises but does not use TempoSync
so I need to fix that. https://github.com/surge-synthesizer/surge/issues/886
is the surge issue for that. Push this anyway since once I fix surge and pull
it will correct this bug.