SurgeModuleCommon data location
Used to be I correctly put surge-data in dist but that all changed
sometime, but I sitll had a stale directory so i didn't know it.
User @SteveRussell33 found the error. This is the obvious fix
which is "dist has build in the name so put build in the name".

Closes #203.