Further cleanup to Rack API shift (#274)
the rack macro INFO went from "logger::log" to "rack::logger::log"
which is a breaking change if you aren't using namespace rack, which
I'm not. And there's no release planned for that change but the source
buidls have it and the SDKs don't. So this fix works through that again
since I keep breaking my code.

Closes #273