Clamp filter frequency bounds
The frequency upper bound used to drive the filter into NaN
state occasionaly which I fixed with an is_infinite in 1.beta1.2
but that didn't stop the filter from blowing up. This change sets
a bound on the filter frequency in both the knob and knob + CV
dimensions to make sure we don't go outside the stable range.

Closes #275