Setup scene/global data in init stage (#480)
My theory for the clasic crash in #462 is that our params are
good but our scenedata isn't when we init. Note that the line
which is crashing (230 in classic) refers to the localcopy
on the get extended and that won't be initialized without this,
so I think it is really it.