Change modulation polyphony when mismatched (#581)
If your module has polyphony N and your modulator M and
M != N we used to always use channel 1. This made sense
when M == 1 but had the counter-intuitive behavior of
making poly signals implicitly mono silently.

So now in the case of M != N and M != 1, we instead use
as much of M as we can and pad the rest with 0. So to be clear
if you ahve a polyphony 5 module and polyphony 3 modulator,
the modulation amounts will be (ch1, ch2, ch3, 0, 0) whereas
before this change they were (ch1, ch1, ch1, ch1, ch1), but
that's really unintuitive it turns out.