First step towards Wavetable Content Download
First step in 2.1 of moving the wavetables mostly out of the
plugin and into an extra content download.

This doesn't change the .vcvplugin but instead makes a target and
action which uploads extra content from head on push. We probably want
a version strategy later for this but for now lets see if this works.

Addresses #710