Updare VCO Modulation Tooltip Display
Modulation tooltip was based on oscstorage_display which was modulated
to animate the display so it was totally wacky. Change it so the display
has the base values and modify through the tp[] structure etc... so the animation
continues but the tooltips reflect base properly. Closes #636

Also set defaults for rate on QuadLFO propelry and fix a typo there in a dropdown