Move to basic-blocks; add an EGxVCA Slow Mode
In preparation for SCXT, Start moving some of the shared
code into sst-basic-blocks. The first to go was these envelopes
and some of the DSP we ahve in rack. While at it, use that change
to factor so I can add an EGxVCA slow mode, which addresses #832

Still a bit of refactoring work to do here to split the ADSRDAHD into
an ADSR and DAHD while still sharing code. MOre to come there tomorrow.