Un-shadow Module and ParamID in ParamStepped
ParamStepped used to be not a param widget but now it is; but
we still had the apocryphal shadows of module and paramid.
This was fine-ish except when you asked for widget->getParamQuantity
in a context where you didn't check for NULL. Stormelder strip
is such a context so this would crash when strip randomized.

Closes #876